requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
chromadb
openai>=1.30.0
tiktoken